pvr: Fix size of flags variable
The poly flags are 16 bits, so the variables should be defined as 16
bits and not 8 bits. This works before because all the flags set in that
particular code were in the first 8 bits, but that was quite dangerous
nonetheless.

+++ src/pvr.c
@@ -2729,7 +2729,7 @@ static void process_gpu_commands(void)
 			uint16_t w, h, x0, y0, x1, y1;
 			bool bright = false;
 			uint32_t color;
-			uint8_t flags = POLY_4VERTEX;
+			uint16_t flags = POLY_4VERTEX;
 
 			if (!textured || !raw_tex) {
 				/* BGR->RGB swap */
